requirements.txt
beautifulsoup4==4.12.2
python-dotenv==1.0.0
openai==1.23.2
streamlit==1.34.0